Initial Website
Hi all. First Commit. Revamped the website, added in goodies, and so on.
Moving TItle, Support for most browsers, links to other resources, and so on. If the moving title isn't welcome, just delete the title.js file and change the title in index.html. Credit to @Bankst for the favicon.