Initial Website (#1)
* Initial Website

Hi all. First Commit. Revamped the website, added in goodies, and so on.
Moving TItle, Support for most browsers, links to other resources, and so on. If the moving title isn't welcome, just delete the title.js file and change the title in index.html. Credit to @Bankst for the favicon.

* Favicon Updating + Getting rid of needless images.

Added 16 and 32 bit favicons, made the 32 bit one the standard (I think). Got rid of extra images since the Logos PNG files are now a part of the website.

* [Main Page] Branding Fix

Fixed the naming to match branding.

* [Colours] Matched Branding

Changed the colours to match the official branding colours in https://github.com/PhotonVision/photonvision-branding/blob/master/PhotonVision%20Branding/Documentation/Photon%20Vision%20Branding%20Standards.pdf

* [Colours] Background

Matched body background with main background.

* [Header] Title/Image

Commented out the picture for the header title, instead made it the "The Future is in Sight" tagline. Picture commented out in case someone makes tagline picture in the future, or if the community wants to switch back.

* [Header] Left Side temp change

* [Landing] Button Restructure

Made the button locations a bit better, colours easier on the eyes.

* [Header]

Fixed text to make it clearer and simpler.

* [Title] Reverted Title

Reverted the title back to static, fixed GitHub name.

* [Landing] Fixes and Features

Fixed the arrow, added 2 features to replace the placeholders.

* [Header] Left Header Title

Need to work on this later, unless someone else is willing to.

* [Update]

Recent import from d4rkf0x.github.io. Everything there applied.

* [Downloads] Typo Fix

Typo

* [General] Update to current Github Pages website

Thanks to Modelmat for updating the icons to FontAwesome5, and changing them to different icons based on votes. Man did me a solid by doing so

* [Landing] Color Update

I'm not sure why this didn't save earlier